rename: useGetUsers의 디렉토리 구조를 변경
